Disable the usuage of the multihop source map for the NBIP lab
The multihop source map needs to be disabled for the NBIP lab, because otherwise the loopback address of the router is used for the eBGP multihop session. This won't work, the eBGP address should be used. This flag disables the usage of the loopback address, and thus the eBGP source address will be used.