Remove AS56647
We don't have any IX in common with FusionMediaLimited (AS56647) in common anymore. They didn't clean up their PeeringDB record, so the sessions were still in place. The AMS-IX notified us that we were hitting the arp sponge at their end. The Speed-IX (the only other IX we had in common) IP didn't respond as well, so the decision was made to remove the peering sessions at our end